Move users/ to authors/ and /u links to /a.
